+++ live-eval/live-eval.go
@@ -25,19 +25,13 @@ const (
 // will return 2 stases (color for Runtime / Dev.) and the list of deps with
 // their color.
 func LiveEvaluation(files []string) error {
-	// Create an array with files content
-	depFiles := make([]models.DependencyFile, len(files))
-	for i, file := range files {
-		depFile := models.DependencyFile{Path: file}
-		content, err := ioutil.ReadFile(file)
-		if err != nil {
-			return err
-		}
-		depFile.Content = content
-		depFiles[i] = depFile
+
+	dfiles, err := models.LookupDependencyFiles(files)
+	if err != nil {
+		return err
 	}
 
-	requestDeps := map[string][]models.DependencyFile{"dependency_files": depFiles}
+	requestDeps := map[string][]*models.DependencyFile{"dependency_files": dfiles}
 	var jsonResp map[string]interface{}
 
 	opts := &gemnasium.APIRequestOptions{
@@ -46,7 +40,7 @@ func LiveEvaluation(files []string) error {
 		Body:   requestDeps,
 		Result: &jsonResp,
 	}
-	err := gemnasium.APIRequest(opts)
+	err = gemnasium.APIRequest(opts)
 	if err != nil {
 		return err
 	}